+ fix Python 2.7 deprecation warning
git-svn-id: https://free-cad.svn.sourceforge.net/svnroot/free-cad/trunk@5055 e8eeb9e2-ec13-0410-a4a9-efa5cf37419d

+++ src/Mod/PartDesign/Scripts/RadialCopy.py
@@ -14,7 +14,7 @@ from FreeCAD import Base
 def makeCopy(shape, radius, angle):
     mat = Base.Matrix()
     mat.rotateZ(math.radians(angle))
-    step = 360.0 / angle
+    step = int(360.0 / angle)
     shape = shape.copy()
     shape.translate((radius, 0, 0))
     comp = shape.copy()
